+++ flake.nix
@@ -1,0 +1,25 @@
+{
+  description = "Platoon development environment";
+
+  inputs = {
+    nixpkgs.url = "github:nixos/nixpkgs?ref=nixos-unstable";
+  };
+
+  outputs =
+    { nixpkgs, ... }:
+    let
+      system = "x86_64-linux";
+    in
+    {
+      devShells."${system}".default =
+        let
+          pkgs = import nixpkgs { inherit system; };
+        in
+        pkgs.mkShell {
+          hardeningDisable = [ "fortify" ];
+          packages = with pkgs; [
+            go
+          ];
+        };
+    };
+}